+++ console/console-init/ui/src/modules/device/utils/constant.ts
@@ -1,0 +1,63 @@
+import { ISelectOption } from "utils";
+import { IDeviceFilter } from "../components";
+
+const deviceTypeOptions: ISelectOption[] = [
+  {
+    key: "direct",
+    value: "direct",
+    label: "Directly connected"
+  },
+  {
+    key: "gateway",
+    value: "gateway",
+    label: "Using gateways"
+  },
+  {
+    key: "alltypes",
+    value: "allTypes",
+    label: "All types"
+  }
+];
+const deviceStatusOptions: ISelectOption[] = [
+  {
+    key: "enabled",
+    value: "enabled",
+    label: "Enabled"
+  },
+  {
+    key: "disabled",
+    value: "disabled",
+    label: "Disabled"
+  },
+  {
+    key: "allstatus",
+    value: "allStatus",
+    label: "All status"
+  }
+];
+
+const getInitialFilter = () => {
+  let filter: IDeviceFilter = {
+    deviceId: "",
+    deviceType: "allTypes",
+    status: "allStatus",
+    filterCriteria: [],
+    addedDate: {
+      startDate: "",
+      endDate: ""
+    },
+    lastSeen: {
+      startTime: {
+        form: "hr",
+        time: ""
+      },
+      endTime: {
+        form: "hr",
+        time: ""
+      }
+    }
+  };
+  return filter;
+};
+
+export { deviceTypeOptions, deviceStatusOptions, getInitialFilter };
