+++ console/console-init/ui/src/modules/device/utils/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from "./TableFormatter";
+export * from "./constant";
